ligify/requirements.txt
requests
xmltodict
biopython
dnachisel
python-dotenv
marshmallow
numpy
boto3
lxml